requirements.txt
streamlit
gunicorn
pandas
supabase
python-dotenv
requests
hmac
hashlib